fups counter + little camera wiggle
since i'm not outputting to a static image, I thought it might be fun to
render things "live", and see how many frames per second this thing can
push. It'll be interesting to see how the number changes as more
features are added.

Oh, and I pass along the time since start to the render method, which
can be used to add a little bit of camera wiggle between frames :D