chapter 8: multiple m a t e r i a l s oh boy
I've never actually worked this much with Trait Objects before in Rust,
it's quite an interesting construct.

I've glanced at some other repos, and it looks like others have taken a
Enum-based approach to the various Hittable and Material objects. This
approach is effectively the "have a single Material type with some of
the fields zeroed out" in terms of space and dispatch (albeit behind a
prettier interface of being an Enum). My gut says that this approach
might be faster than dynamic dispatch, but on the other hand, it wastes
quite a bit of space, and isn't easily extensible by external users
(i.e: adding a new material requires changing the ray-tracer's core
code, as opposed to implementing your own type that implements Hittable,
and passing it as a Box<dyn Hittable> to the existing code...

It might be interesting to experiment with the two techniques once the
ray tracer is further along, and profile the two against eachother.