explore animating the world
At the moment, I keep the spheres vector as a right-proper Vec<Sphere>,
and on every loop, I regenerate the world Vec<&dyn Hittable> from the
spheres.

This works _fine_, but really, it'd be great to be able to "cache" this
alternative view of the spheres (i.e: as Trait Objects).

I actually went on a _really_ long tangent exploring different ways to
accomplish this, including some forays into unsafe territory, and trying
to use the downcast-rs crate -- all to no avail...

Anyways, it's not mission critical or anything. It's moreso been an
interesting exploration into bending the borrow checker to my will (and
kinda failing to do so lol).