chapter 12: r a n d o m balls
and that's that then! that's the book!
I'll definately be following along with the next book as well, but
before doing so, I'll clean up this current repo some more.

A couple of items I want to get done:
- multithreading!
- separating library + driver code
- experimenting with enum-dispatch instead of dynamic dispatch
  (for performance)
- general cleanup and refactoring
- more that I can't think of right now lol

I did a bit of refactoring around defining scenes, breaking them off
into their own submodule.