parallelism! thanks rayon!
I had to sprinkle some `Sync` traits here and there, but otherwise,
making this thing fire on all cylinders was fairly unobtrusive!

Next up, I want to refactor the code to allow a live-updating
framebuffer as pixels are rendered on screen. Instead of just a
terminal-based progress bar, I was to see the image as it gradually
appears on screen!