add toggle between static and dynamic dispatch
wew what a change.

I'm pretty happy with how the changes ended up. I was expecting a lot
worse. The conditional compilation flags are very well localized, with
the rest of the codebase responding transparently when the underlying
dispatch mode is changed.

As far as I can tell, there isn't too much of a performance difference
between the two dispatch types. This could be because there aren't
enough objects in the scene to make a big difference, or maybe there
aren't enough variants for the two modes to really shine, who knows?
Heck, maybe I'm a Bad Programmer, and my flags aren't even working haha
(though I doubt that).

Anyhoo, it'll be interesting to profile the two dispatch types as the
raytracer gains more functionality. Also, It'll be interesting to see if
I can maintain this level of abstraction as more unforseen functionality
is added to the codebase haha.